fix build
used int instead of unsigned
strange that no trybot caught it

TBR=reed@google.com

Bug: skia:6898

+++ tools/bookmaker/bookmaker.h
@@ -2367,7 +2367,7 @@ public:
         : TextParser(fileName, start, end, lineCount)
         , fClassName(className)
         , fLocalName(className) {
-        int doubleColons = className.find_last_of("::");
+        size_t doubleColons = className.find_last_of("::");
         if (string::npos != doubleColons) {
             fLocalName = className.substr(doubleColons + 1);
             SkASSERT(fLocalName.length() > 0);
